+++ BlackDragon.cpp
@@ -11,14 +11,15 @@ BlackDragon::BlackDragon():Creature(9,10,200,blackDragon){}
 //NO SPECIAL
 void BlackDragon::specialAbility(Creature &c){return;}
 
-int BlackDragon::attackAnother(Creature &c)
-{
+double BlackDragon::attackAnother(Creature &c){
     cout << "BlackDragon go for attack" << endl;
-    int ratio;
+    double ratio;
     if (c.getType() == 3) c.specialAbility(*this);
     ratio = getAttackPoints() / c.getDefendPoints();
     return ratio;
 
 }
+
 CreatureType BlackDragon::getType(){return blackDragon;}
+
 BlackDragon::~BlackDragon(){}
