+++ Zombie.h
@@ -19,7 +19,7 @@ public:
 
 	virtual ~Zombie();
 
-	int attackAnother(Creature &c);
+	double attackAnother(Creature &c);
 	void specialAbility(Creature &c);
 	int getCost(){return 50;}
 	string displayType(){return "Zombie";}
